Handling update score action
Added a edit button on every item of summary to enable the edit mode.
Created a EditGameItem to display handle the form submit and dispatch the update score action.
Disables the edit and finish button when game item is on edit mode.
Added a cancel button to exits the edit mode without changes.
Added tests to validate the basic edit functionality, checks the summary item updated.